+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Bug #67: Fix hardcoded items children table name in item tree traversal query for MySQL 5 (@arogachev)
 - Enh #71: Improve perfomaтce (@arogachev)
 - Chg #78: Raise PHP version to 8.1 (@arogachev)
+- Chg #79: Add customizable separator for joining and splitting item names (@arogachev)
 
 ## 2.0.0 April 20, 2023